# Fix performance of broadcast and collect with Union{T, Missing}