Use the same pattern as in `collect_to!` (which is used when size is known). Follows what was done by https://github.com/JuliaLang/julia/pull/25828.

Fixes https://github.com/JuliaLang/julia/issues/30455.

Before:
```julia
julia> using BenchmarkTools

julia> v = fill(2.0, 50000);

julia> v2 = vcat(v, missing);

julia> v3 = vcat(missing, v);

julia> @btime $v .* 3.0;
  46.964 μs (2 allocations: 390.70 KiB)

julia> @btime $v2 .* 3.0;
  1.838 ms (99498 allocations: 2.33 MiB)

julia> @btime $v3 .* 3.0;
  3.569 ms (7 allocations: 439.77 KiB)

julia> @btime collect($(x for x in v if true));
  336.170 μs (17 allocations: 1.00 MiB)

julia> @btime collect($(x for x in v2 if true));
  569.995 μs (21 allocations: 2.29 MiB)

julia> @btime collect($(x for x in v3 if true));
  4.041 ms (20 allocations: 1.13 MiB)

```
After:
```julia
julia> @btime $v .* 3.0;
  46.908 μs (2 allocations: 390.70 KiB)

julia> @btime $v2 .* 3.0;
  1.351 ms (99498 allocations: 2.33 MiB)

julia> @btime $v3 .* 3.0;
  102.446 μs (7 allocations: 439.77 KiB)

julia> @btime collect($(x for x in v if true));
  332.648 μs (17 allocations: 1.00 MiB)

julia> @btime collect($(x for x in v2 if true));
  527.674 μs (21 allocations: 2.29 MiB)

julia> @btime collect($(x for x in v3 if true));
  625.985 μs (19 allocations: 1.13 MiB)
```