@nalimilan 2018-12-21
@nanosoldier `runbenchmarks("array" || "broadcast" || "union", vs = ":master")`