@nalimilan 2018-12-21
Nanosoldier confirms this improves performance dramatically with `Union{T, Missing}`. Most regressions are clearly spurious, there's only one regarding array comprehensions which could be real (but that would be surprising since there's no type instability): is it known to be noisy?