@ararslan 2018-12-21
@nanosoldier `runbenchmarks("array" && "comprehension", vs=":master")`