@nanosoldier 2018-12-21
[Your benchmark job](https://github.com/JuliaLang/julia/pull/30480#issuecomment-449480937) has completed - no performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/1c681ee273d698f7531ed7dcddd605ce62c70343/16b0905_vs_163b019/report.md). cc @ararslan