@nalimilan 2018-12-21
> Am I understanding the magic here correctly? Is the key that we should not assign a type that may be a small discriminated union to a variable?

That's my conclusion as well, but I have no idea why that's the case.