@vtjnash 2018-12-21
`<:` has almost no optimizations implemented (there aren’t many worth doing) while `isa` is heavily optimized and usually gets replaced by something very simple